Extract `useDomains` hook so it can be reused on `AddEmailForwarder`
